requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
pydantic==2.8.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
spacy==3.7.4
pdfplumber==0.10.3
python-docx==0.8.11
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==9.5.0
gspread==5.10.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
twilio==8.10.0
aiofiles==23.2.1
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1
faker==19.6.2
black==23.9.1
isort==5.12.0
ruff==0.0.287
mypy==1.5.1
